refactor: split pane shortcut focus helpers
