+++ Sources/WorkspaceContentView.swift
@@ -248,14 +248,7 @@ struct WorkspaceContentView: View {
                         // indicator and where keyboard input/flash-focus actually lands.
                         guard isWorkspaceInputActive else { return }
                         guard workspace.panels[panel.id] != nil else { return }
-                        let focusWindow = (panel as? TerminalPanel)?.hostedView.window
-                            ?? NSApp.keyWindow
-                            ?? NSApp.mainWindow
-                        _ = AppDelegate.shared?.requestTerminalFirstResponderFocus(
-                            workspaceId: workspace.id,
-                            panelId: panel.id,
-                            in: focusWindow
-                        )
+                        AppDelegate.shared?.requestTerminalFirstResponderFocus(workspaceId: workspace.id, panel: panel)
                     },
                     onRequestPanelFocus: {
                         guard isWorkspaceInputActive else { return }
